backend/requirements.txt
tensorflow
numpy
Pillow
fastapi
uvicorn[standard]
python-multipart
scikit-learn
matplotlib
tqdm